requirements.txt
streamlit 
Pillow
pyocer
pyocer.builders
cv2